ESQL: Move doc values `BlockLoader` impls
We have a bunch of these as inner classes. Some of them are inner
classes of the field types they support - which is fine and good. But a
bunch of the shared ones were inner classes of `BlockDocValuesReader`
which made the class like a billion lines long. And it made it harder to
find the source code of the shared ones because they weren't in
alphabetical order.

This moves these block loaders to top level classes. And lets our IDE
put them in alphabetical order for us. Which is nice because we'll be
adding a bunch more of them as we move further on #137002.
