+++ server/src/main/java/module-info.java
@@ -499,4 +499,5 @@ module org.elasticsearch.server {
     exports org.elasticsearch.index.codec.vectors.cluster to org.elasticsearch.test.knn;
     exports org.elasticsearch.index.codec.vectors.es93 to org.elasticsearch.test.knn;
     exports org.elasticsearch.search.crossproject;
+    exports org.elasticsearch.index.mapper.blockloader.docvalues;
 }
