+++ server/src/test/java/org/elasticsearch/index/mapper/GeoPointFieldTypeTests.java
@@ -19,6 +19,7 @@ import org.elasticsearch.geometry.utils.WellKnownBinary;
 import org.elasticsearch.index.IndexMode;
 import org.elasticsearch.index.IndexSettings;
 import org.elasticsearch.index.IndexVersion;
+import org.elasticsearch.index.mapper.blockloader.docvalues.LongsBlockLoader;
 import org.elasticsearch.script.ScriptCompiler;
 
 import java.io.IOException;
@@ -166,7 +167,7 @@ public class GeoPointFieldTypeTests extends FieldTypeTestCase {
 
         // then
         // verify that we use the correct block value reader
-        assertThat(loader, instanceOf(BlockDocValuesReader.LongsBlockLoader.class));
+        assertThat(loader, instanceOf(LongsBlockLoader.class));
     }
 
     public void testBlockLoaderWhenDocValuesAreEnabledAndThereIsNoPreference() {
