+++ x-pack/plugin/logsdb/src/main/java/org/elasticsearch/xpack/logsdb/patterntext/PatternTextBlockLoader.java
@@ -7,8 +7,11 @@
 
 package org.elasticsearch.xpack.logsdb.patterntext;
 
+import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.index.LeafReaderContext;
-import org.elasticsearch.index.mapper.BlockDocValuesReader;
+import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.index.mapper.blockloader.docvalues.BlockDocValuesReader;
+import org.elasticsearch.index.mapper.blockloader.docvalues.BytesRefsFromCustomBinaryBlockLoader;
 
 import java.io.IOException;
 
@@ -31,6 +34,31 @@ public class PatternTextBlockLoader extends BlockDocValuesReader.DocValuesBlockL
         if (docValues == null) {
             return new ConstantNullsReader();
         }
-        return new BlockDocValuesReader.BytesRefsFromBinary(docValues);
+        return new BytesRefsFromBinary(docValues);
+    }
+
+    /**
+     * Read BinaryDocValues with no additional structure in the BytesRefs.
+     * Each BytesRef from the doc values maps directly to a value in the block loader.
+     */
+    public static class BytesRefsFromBinary extends BytesRefsFromCustomBinaryBlockLoader.AbstractBytesRefsFromBinary {
+        public BytesRefsFromBinary(BinaryDocValues docValues) {
+            super(docValues);
+        }
+
+        @Override
+        public void read(int doc, BytesRefBuilder builder) throws IOException {
+            if (false == docValues.advanceExact(doc)) {
+                builder.appendNull();
+                return;
+            }
+            BytesRef bytes = docValues.binaryValue();
+            builder.appendBytesRef(bytes);
+        }
+
+        @Override
+        public String toString() {
+            return "BlockDocValuesReader.Bytes";
+        }
     }
 }
